requirements.txt
PySide6==6.6.2
python-dotenv==1.0.1